pyproject.toml
[project]
name = "sprintboy"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "aiohttp>=3.13.1",
    "langchain>=1.0.2",
    "langchain-anthropic>=1.0.0",
    "pydantic>=2.12.3",
    "pytest>=8.4.2",
    "pytest-asyncio>=1.2.0",
    "python-dotenv>=1.1.1",
]

[dependency-groups]
dev = [
    "black>=25.9.0",
    "cyksuid>=2.1.0",
    "ipython>=9.6.0",
    "polars>=1.34.0",
    "pyright>=1.1.407",
    "ruff>=0.14.2",
]
